+++ src/main/java/com/example/homeaid/customer/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
 package com.example.homeaid.customer.repository;
 
-import com.example.homeaid.customer.entity.Customer;
+import com.example.homeaid.customer.customer.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
