+++ src/main/java/com/example/homeaid/manager/repository/ManagerRepository.java
@@ -1,6 +1,6 @@
 package com.example.homeaid.manager.repository;
 
-import com.example.homeaid.manager.entity.Manager;
+import com.example.homeaid.manager.manager.entity.Manager;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ManagerRepository extends JpaRepository<Manager, Long> {
